// cx-organization-spa/src/app/user-accounts/models/edit-user.model.ts
export class EditUser {
  firstName: string;
  identityType: string;
  ssn: string;
  emailAddress: string;
  dateOfBirth: string;
  expirationDate: string;
  gender: number;
  activeDate: string;
  teachingRecords: string;
  personalStorage: number;
  isStorageUnlimited: boolean;
  entityStatus: string;
  personnelGroups: any;
  departmentId: number;
  designation: string;
  portfolio: any[];
  jobKey: string;
  signupReason: string;
  careerPaths: any[];
  teachingLevels: any[];
  teachingCourseOfStudy: any[];
  systemRoles: any[];
  teachingSubjects: any[];
  cocurricularActivities: any[];
  roleSpecificProficiencies: any[];
  jobFamily: any[];
  developmentalRoles: any;
  learningFrameworks: any[];
  professionalInterestArea: any[];
  constructor(data?: Partial<EditUser>) {
    this.firstName = data.firstName;
    this.identityType = data.identityType;
    this.ssn = data.ssn;
    this.emailAddress = data.emailAddress;
    this.dateOfBirth = data.dateOfBirth;
    this.expirationDate = data.expirationDate;
    this.gender = data.gender;
    this.activeDate = data.activeDate;
    this.teachingRecords = data.teachingRecords;
    this.personalStorage = data.personalStorage;
    this.isStorageUnlimited = data.isStorageUnlimited;
    this.entityStatus = data.entityStatus;
    this.personnelGroups = data.personnelGroups;
    this.departmentId = data.departmentId;
    this.designation = data.designation;
    this.portfolio = data.portfolio;
    this.jobKey = data.jobKey;
    this.signupReason = data.signupReason;
    this.careerPaths = data.careerPaths;
    this.teachingLevels = data.teachingLevels;
    this.teachingCourseOfStudy = data.teachingCourseOfStudy;
    this.systemRoles = data.systemRoles;
    this.teachingSubjects = data.teachingSubjects;
    this.cocurricularActivities = data.cocurricularActivities;
    this.roleSpecificProficiencies = data.roleSpecificProficiencies;
    this.jobFamily = data.jobFamily;
    this.developmentalRoles = data.developmentalRoles;
    this.learningFrameworks = data.learningFrameworks;
    this.professionalInterestArea = data.professionalInterestArea;
  }
}